[flow] Add optional use_desc to get_builtin_typeapp
Summary:
Now we add an optional use_desc arg to get_builtin_typeapp. It may be better to make this required, but it is not on `get_builtin_type`.

This is the final diff that sets the stage to use RReactElement as a reason desc for component syntax elements. The next diff makes the desc change.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D49423115

fbshipit-source-id: d7f8fe51d6b2547659529397456bb5151a12f871