[flow] Add use_desc to typeapp
Summary:
TypeAppTs, when evaluated during subtyping, are always repositioned to use the [reason for the tapp](https://github.com/facebook/flow/blob/main/src/typing/subtyping_kit.ml#L678-L702), but they use the reason `desc` from the target of the type applications.

This makes it hard to control the user-visible reason used for a type application, as the desc is always taken from the target. In particular, that makes it difficult to [emit good error messages for `React.Element`](https://github.com/facebook/flow/blob/main/src/typing/react_kit.ml#L664-L667). All React.Elements will use a reason that prints as `React.Element`, but we could instead use the `RReactElement of name` reason desc, which prints as `name element`, when we are instantiating a component-syntax component. Changing the reason_desc is not sufficient, as it will get overridden by the TypeAppT's repositioning. Repositioning the result of the `get_builtin_typeapp` call does not work, as the repositioning is pushed to the unevaluated TypeAppT, which again gets overridden when evaluated. There is no reasonable way to guarantee the right reason desc is used for an unevaluated typeapp.

This diff adds a `use_desc` field to TypeAppT, which we can use to set the `use_desc` fields to true in the various repositioning functions called by the TypeAppT evaluations.

This diff does not make use of the new field and only gets the system to build. I also changed the TypeAppT payload to an inline record with this change for better documentation.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D49420939

fbshipit-source-id: bcdcc70012200f4383ff4da806fc38da8d302a56