[flow] Require use_desc arg on typeapp creation functions in TypeUtils
Summary:
Building on top of the previous diff, let's make all callers to `typeapp` manually specify the `use_desc` argument. This makes it clear that it is an important part of the public API of a typeapp and you should choose true or false with care.

We still do not set this to true anywhere, but we thread arguments as necessary

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D49422836

fbshipit-source-id: ab4e3218a1836740ce9249c413ad3505ee923087