[flow] specialized callee types in TAST
Summary:
Before, hover type on
```
declare function getElement<X>(x: Array<X>): X;
getElement(["abc"]);
// ^
```
would return `<X>(x: Array<X>) => X`, instead of the specialized form based on the arguments to the call `(x: Array<string>) => string`.

Similarly, hovering over the callee at a function call to an intersection type would show all branches of the intersection, instead of the one selected.

This diff fixes these cases, by storing a specialized form of the type of a callee in the typed AST, so that it can be used by all type services (e.g. hover types).

Implementation wise, we store the necessary information in mutable fields under various use types that are associated with calls, e.g. CallT, MethodT. The information we store includes
* a (list of) "finalized" type(s) that hit the `l ~> CallT _` constraint. The reason we need a list here is the case where `l` is a union, which first gets broken up into its members. These type populate the `finalized` field. We record types in this field when we're not operating under speculation.
* a list of "speculative candidates" that are accumulated as we perform the call for the various branches of an intersection. We also record the speculative state under which this addition was made. After speculation is done (at the "fire action" part of Speculation_kit) we determine if the branch was successful, and if so, we promote the type to the finalized list.

This diff also cleans up a utility that was earlier used to record the (uninstantiated) type of the receiver at a method call. This was the the last element of the MethodT and other constructors. This type probe is no longer necessary for most cases. It is only useful for the `NoMethodAction` case, and so it is converted to a payload to the `NoMethodAction` constructor.

These field under `specialized_callee` are updated appropriately in a few places. These places correspond roughly to the places where we earlier updated the MethodT receiver probe. This is mainly `apply_method_action` in flow_js.ml, but there's a couple more interesting places:
* the fall-through case of the Flow_js pattern match. This is handled in default_resolve.ml.
* Speculation_kit. We handle both the case of a successful branch, where we finalize the case that won, and the case of failure, where we need to reconstruct the intersection type that caused speculation in the first place.

Changelog: [fix] hover-types will now show a specialized version of the callee expression at a call to a generic or intersection-typed function

Reviewed By: SamChou19815

Differential Revision: D47180573

fbshipit-source-id: b41e3b43c9539eb6b7c0242b23bde5b7587dc472