[flow] Make array and object spreads apply DRO to their elements but not be DRO in the result
Summary: A typical pattern, and one we want to support, for cases where components want to pass on a modified version of their props to other components, is to create a copy of the props value with an array or object spread. Previously, this copied value was still seen as DRO, which would be a problem because it would discourage the correct behavior. This diff makes spread copies non-DRO, but it also ensures that the members/elements of the copied value are DRO, because spreading creates a shallow copy, not a deep one.

Reviewed By: SamChou19815

Differential Revision:
D49274413

------------------------------------------------------------------------
(from 2c75b6eb8283ab28d545384b90193204d586de97)

fbshipit-source-id: 8f72490d53ccab2b0bc2bae2ac614619a1713258