[flow][perf] Record an optional source aloc id as a stable identity for unions for fast path union check
Summary:
This optimization is inspired by the idea that we use the id of object props for optimization purposes:

https://github.com/facebook/flow/blob/6a6f0786e3dfeedd1d52457ba14e65ada414ca28/src/typing/subtyping_kit.ml#L1636-L1638

Can we do the same for union types? We can, and this is exactly what this diff does.

For union types that come from annotations, we record the source location id (aloc) into the union type. When doing a fast-path check for two unions, we first check whether the source id are the same. Previously, we do a full structural check of two unions, but this optimization can be easily destroyed after union type repositioning. To keep this, I ensure that repositioning does not change the tracked source id, because repositioning does not change the semantics of type.

Can we generalize the idea to associate every union with an id? We could, and I tried it in D49438687. I have observed that it results in even greater perf improvements, but it breaks some of the typeapp termination guard and causes some weird errors. That's because it can result in two union types that are structurally the same, but a structural comparison or equality check in ocaml will consider them to be different, because they have different ids. To correct that, we need to either do some union type interning (which is not guaranteed to help perf), or change all of the `TypeMap` and `TypeSet` to be aware of this, which is quite invasive that I don't have time to do yet.

Changelog: [perf] We slightly optimized Flow's performance of subtyping check between two union types.

Reviewed By: panagosg7

Differential Revision: D49475304

fbshipit-source-id: abf7e42864a01951f183d5b114ddcb3e014ce10e