[flow][render-types] Initial cut for render types argument validation
Summary:
This diff implements an initial version of render types validation behind a flag. For now, we will accept only unions of `React.Element<an nominal component>`, or an `GenericT` under component type or declaration. (I think it can be any GenericT, since we already validated that it's a subtype of `React.Node` elsewhere)

The error messages are not great yet, and there is no autofixes, but that will be the job of followup diffs to fix them.

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D49292399

fbshipit-source-id: dfa134ceb70ada9ed0b96ee61d8010eb45ca206f