[flow] Prepare renders normalization and subtyping rules for variants
Summary:
Our render type normalization will squash nested `RendersT`s, which will no longer work when we have mulitple variants of RendersT. The reason we do that is to make the Nominal ~> Structural subtyping case a bit easier-- we examine the payload of the Structural side for an id, and if it is no present we recurse to the super.

This check is attempting to be a performant way to avoid having to do a speculative subtyping check of the nominal LHS against the type on the RHS, but this check will no longer work with additional render variants.

This diff implements the rule in its most basic form instead, which generalizes easily to the other render variants. I'm not very worried about making this check as performant as possible because the render type validation we introduce will push usage into the very performant cases only.

In total, this diff:
1. Gets rid of the extra normalization
2. Changes the Nominal ~> Structural subtyping rule into a more general form that will work for all renders variants

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D49506733

fbshipit-source-id: 9ee5dcbdd1c60ced3ae8cd8ec8cf910a728eee8c