[flow] Introduce renders_variant for render types
Summary:
Add a `renders_variant` field to canonical_render_form that indicates if the render type is a `renders` `renders?` or `renders*`.

This diff introduces the `RendersNormal` variant, which is what the existing subtyping rules support, and fixes all existing usages of render types to build. This diff adds no new behavior.

RendersMaybe and RendersStar are stacked on this diff.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D49513703

fbshipit-source-id: 491247a2cd8eb8e5ac5fcbbf28d911456310fe05