[flow] Always use long-lived worker mode under Windows
Summary:
We recently implemented and rolled out long-lived workers on Windows. Before
long-lived workers, Flow would spawn a process for each parallel job. This had
terrible performance, and long-lived workers was a huge perf win.

This diff is a bit of a shortcut. Instead of making LLW the default, then
eventually removing the old behavior, I'm jumping right to the end. This diff
does preserve the flowconfig option, but it does nothing. I will remove the
config in the next version.

Removing the spawn-on-each-job behavior allows some nice clean-up of the code as
well, which previously had to deal with "maybe prespawned" workers, but can now
assume all workers are prespawned.

I also changed the heap tests to use long-lived workers on all platforms. We
will eventually kill forking altogether, at least once we are using OCaml 5
domains, if not much sooner.

Reviewed By: SamChou19815

Differential Revision: D49710582

fbshipit-source-id: 19f648190ca4133ababe3d81e466b70310b0db94