[flow][autoimport] Insert nicer path for haste package auto imports
Summary:
We currently insert a very ugly relative path imports for files inside a haste package. However, we do have special logic for `node_modules` path, and we can do the same for haste.

The logic is actually fairly simple: for a path like `a/b/c/foo/bar/baz.js`, we check whether `bar` is a haste package. If so, the import path should be `bar/baz.js`. Then we check whether `foo` is a haste package, if so, then the import path should be `foo/bar/baz.js`, and so on.

Changelog: [internal] I don't think any external folks use Haste.

Reviewed By: mroch

Differential Revision: D49844763

fbshipit-source-id: 4f5b17e6505c7c8708e6cc81ce3f16290c281fa7