[flow][render-types] Change Component return AST to use a Render Type directly
Summary:
Components will always be annotated with a `renders` type, but just storing the Type that comes after the `renders` token is no longer sufficient with the renders variants. We need to also keep track of which renders variant it is.

A simple way to do this is to just store the Renders.t AST structure directly in the render_type field of a component. Simple in theory, but a large diff to make it happen.

This diff makes that change, including changes to ESTree output which will necessitate prettier updates.

Since these new types are still not parsed (D49929968 implements parsing), I can't test in this diff.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D49759163

fbshipit-source-id: b6cb808af46480ed7ec0fea7b3be1bd0ec2bd390