[flow][render-types] Renders variants parsing
Summary:
This diff adds support for parsing `renders?` and `renders*`.

`renders?` is a bit tricky since it may be a function parameter or a property name. The parser already handles the object property case, but the function parameter case requires some special care. See the added tests for the corner cases.

`renders*` is much more straightforward and I have not thought of any tricky parsing cases for it.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D49929968

fbshipit-source-id: fd4ae011d57daf32369c9fca219452ac88e85678