[flow][render-types] Hover documentation for render types
Summary:
This diff adds hover documentation for render types. Ideally we only do this on the keyword, but I'd like to wait until jbrown215's render type AST work is complete to do that.

We currently have infra for associating jsdoc with renders, but since this is a new hardcoded documentation, I just added another searcher. I also need to change the LSP handling code so that we can still output documentation when we don't have type-at-pos result.

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D49924434

fbshipit-source-id: 2ac2095ddf3596fcf9f672046f5a59daba552b8e