[flow] Add casting_syntax option, allows `as` expression based type casting
Summary:
Changelog: [feature] Adds `casting_syntax` option which can be `colon` (current), `as`, or `both`.

If setting is `both`, both `(x: T)` and `x as T` are allowed. If set to `colon` or `as`, only `(x: T)` and `x as T` are allowed respectively (with quickfix to the other).

Open to suggestions on the option name and values

Reviewed By: SamChou19815

Differential Revision: D49835245

fbshipit-source-id: e759275dc629126b3f049106a84a415a73becd38