[flow][normalization] Print React.Element syntactic sugar
Summary:
We normalizing the type `Foo` given a component `Foo` we should use the shorthand in the normalization result.

We do this by adding a special case to TypeAppT to detect the shorthand syntax based on the type's reason. While reasons are not always reliable, this is a common pattern in the normalizer.

Changelog: [internal]

Reviewed By: SamChou19815, panagosg7

Differential Revision: D49961127

fbshipit-source-id: eeb56c67dc45ee04d403f229379e2553da67801e