[flow] incremental error collation
Summary:
## Motivation

Error collation is the process responsible for
1. determining an error's primary location,
2. determining whether the error should be shown (or be suppressed, and
3. bringing the error to a human readable form.

Collation takes place in the server process after each recheck (assuming at least one open connection to the server, which is typically the case). This means that, while it is fast overall, it lies in the critical path of returning an error report to the user and, more importantly, may block the server from servicing other IDE requests while it's under way.

Until now Flow would compute collated errors using the entire error set that has been accumulated in the server up to that point. On a freshly spun lazy server this set should be small, as are the set of files that are checked, but after a large rebase (or on full inits), it is possible for this set of errors to grow to even a few hundred thousands. In these situations, error collation can take more than 10 sec on *any* recheck even on files with zero fanouts.

## This diff

To mitigate this inefficiency this diff makes the process of collation incremental. In the server environment we now store a structure `incr_collated_errors` that contains maps from file names to concretized error sets. It is important to store error sets this way so that they can be cleared whenever a file is changed, deleted etc.

On rechecks, we update `incr_collated_errors` just on the files of interest. In this iteration of this feature, this will include the file that is changed and all its dependants.

Note that for now we cannot minimize this set based on other recheck optimization we have in place. The reason for this is that even whitespace changes, which typically hit our recheck opts, are important for the stability of concretized errors and so should be taken into account in producing the fresh concretized error set. Consider for example the scenario of tests/abstract_locations_skip_recheck:
* File B depends on file A
* Checking B reveals an error that references a location in file A
* adding whitespace in file A, does *not* cause B to be rechecked. The abstract location in the existing error is still valid, but the concretized view of the reference to A has now changed, so we'd need to recompute the concretized error.

A similar scenario emerges if we consider misplaced errors (errors whose primary location is in a different file that the one that raised them).

There are a couple of ways to ensure the error messages are stable. This diff uses a more coarse grained approach that requires (as stated above) incremental collation to include errors from all files in the dependent set.

An alternative could involve only inspecting errors from the recheck optimized set of dependents, but also keep track of locations in errors coming from the remaining set of dependents in case some of these locations needs to be updated. We will reevaluate the need for such an approach after initial rollout.

Changelog: [perf] error collation during rechecks should become much faster thanks to incremental computation. (Currently feature available behind the flag `--incremental-error-collation`.)

Reviewed By: samwgoldman

Differential Revision: D49844411

fbshipit-source-id: d64e1373733a5beb413bf7944f8500459ace1994