[flow][multi-plat] Allow platform specific files to have different type exports against common interface file
Summary:
In the multiplatform support, we need to check platform specific implementation files against the common interface file. We need to validate all the value exports, but not type exports, since type exports are not used at runtime, you just get what you get.

Until the last diff, there is no reliable way to decide whether something is a type export. This diff uses that to exclude all the type exports when we try to get the signature of a common interface file. Therefore, when we are doing a subtyping of platform specific implementation file against common interface file, we will ignore the differences in type only exports.

Changelog: [internal] multiplatform stuff is still experimental

Reviewed By: samwgoldman

Differential Revision: D50084442

fbshipit-source-id: 716c61bc40587a5bfe20dd261a83879b59eaa2ba