[flow][type-sig] Align type sig treatment of globals with the current globals behavior (1/n): Do not use local definitions for special names
Summary:
The goal is to generate types of globals using type sig. Type sig for globals are not used for typing purposes at all yet, so some behavior has drifted from the existing system.

In the current system of checking libdefs, when we see some special builtin names like `Array`, we still special case its behavior instead of reading from the local scope:

https://github.com/facebook/flow/blob/b4db950d194eaab9e5e9a546bc7b1842eb8a7208/src/typing/type_annotation.ml#L739-L742

This diff changes the behavior in type sig to align with that.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D50291085

fbshipit-source-id: 8b74c26511088bab90d9229104b83419ab8f273e