[flow][type-sig] Align type sig treatment of globals with the current globals behavior (3/n): Always auto export all local types for CJS declare modules
Summary:
The goal is to generate types of globals using type sig. Type sig for globals are not used for typing purposes at all yet, so some behavior has drifted from the existing system.

In the current system of checking libdefs, we always auto export all local types for CJS declare modules:

https://github.com/facebook/flow/blob/b4db950d194eaab9e5e9a546bc7b1842eb8a7208/src/typing/type_env.ml#L870-L901

This diff changes the behavior in type sig to align with that.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D50291003

fbshipit-source-id: b800039b944d2e631b5199b7ed97b8a62473a566