[flow][type-sig] Align type sig treatment of globals with the current globals behavior (2/n): Use name/literal location as reason location of declare module
Summary:
The goal is to generate types of globals using type sig. Type sig for globals are not used for typing purposes at all yet, so some behavior has drifted from the existing system.

In the current system of checking libdefs, we use the location of name/literal as the reason location for declare module:

https://github.com/facebook/flow/blob/b4db950d194eaab9e5e9a546bc7b1842eb8a7208/src/analysis/env_builder/name_def.ml#L2127-L2134

This diff changes the behavior in type sig to align with that.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D50291083

fbshipit-source-id: 392232c72e125fee1c966bad15638eaa3e1eafa7