[flow][type-sig] Align type sig treatment of globals with the current globals behavior (4/n): Special treatment for implicitly exported object for declare module
Summary:
The goal is to generate types of globals using type sig. Type sig for globals are not used for typing purposes at all yet, so some behavior has drifted from the existing system.

In the current system of checking libdefs, implicitly exported stuff from ambiguous declare module will be an object with readonly props, and don't have a object literal reason.

https://github.com/facebook/flow/blob/b4db950d194eaab9e5e9a546bc7b1842eb8a7208/src/typing/type_env.ml#L885-L890

This diff changes the behavior in type sig to align with that. We do not use the `ObjLit` but instead create a new annotation kind that keep track of all the necessary information.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D50291846

fbshipit-source-id: acd3117ffe555d19cb272a09dc6e49adf9c12f24