[flow][tests] Rewrite tests so that it no longer inspects React.Element internals
Summary:
The tests can be rewritten so that they can still test something useful. I rewrite the tests in a way so that instead of testing whether a property of `React.Element` satisfies a certain type, I will test whether a `React.Element` is a subtype of another `React.Element`. This prepares us to make `React.Element` fully opaque in Flow.

Changelog: [internal]

Reviewed By: alexmckenley

Differential Revision: D50340294

fbshipit-source-id: 8ab6170e8ef7a48b4b940d1453da23af9958add1