[flow] add helpful message in error when type guard returns false
Summary:
The type guard consistency check will raise an error in
```
function isString(x:?string): x is string {
  if(x == null){
    return false;
  }
  return true;
}
```
Technically this is not specific to type guards, but refinements in general, since Flow does not use the false assertion (that implies dead code) to refine arbitrary variables. E.g:
```
declare var inputText: ?string;
if (false) {
  (inputText: string); // this is still an error
};
```
In theory, false here is in a position where it could apply any refinement on any variable that is in scope.

This diff improves the message whenever a "return false" statement leads to a consistency error, by including the statement:
```
Consider replacing the body of this predicate function with a single conditional expression.
```
The hope is that this will help the user replace their code with
```
function isString(x:?string): x is string {
  return (x != null);
}
```

Fixing this properly to allow the refinement would be more tricky, due to type guard support for negation at the call site. In the case of "return false" we'd need to prove that the complement of the type of the input refined with existing refinements up to that point and the annotated guard type do not overlap (their value intersection is empty). We'd need this to avoid the unsoundness of code like
```
function pred(x: number | string): x is number {
  return false;
}
// if we naively accept the above definition on the grounds that
// the condition will always fail, then the cast below would succeed
// with no errors

const x: number | string = 1;
if (!pred(x)) {
  (x: string); // okay
}
```

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D50249659

fbshipit-source-id: 5268124c10f3bdedd041b2a995d9ceb8b220a70e