[flow][dev-tools] Use hermes-parser for inline snapshot recording instead
Summary:
I want to get rid of dependency of flow-parser on our dev tools, so that the build setup can be simpler. In this diff, I killed one of the last two usage of flow-parser: recording of inline snapshot.

Currently, we use flow-parser to parse the AST to decide where to replace the inline snapshot in LSP tests. It can be easily done by hermes-parser.

Changelog: [internal]

Reviewed By: pieterv

Differential Revision: D50416330

fbshipit-source-id: 27221b137501e18e729ab5f39ff03179035495df