[flow] Generate libdef types using type sig so that libdefs can be validated under LTI
Summary:
Currently, libdefs cannot be checked under LTI, because we check libdef files one by one, and a file might reference a global defined in a not-yet-visited file, so we have to keep almost all the tvars open.

This diff changes that. We will first using type sig to generate libdefs. The previous diffs even out the differences between the current system and the new system, so that we have almost the same behavior with the generated libdef types.

When validating the libdefs, we already have all the libdef types added to globals. We will validate libdefs under a special mode where the reads of globals in libdef files will not read the local writes, but the already fully resolved globals. In this way, everything read will be fully resolved, so we can check it under LTI.

Under this new system, quite a lot of ugly hacks can be deleted. We can delete all the checks of whether we are in InitLib to skip some LTI checks and tvar resolution. We can also delete some code used to generate env entries, since the types of globals no longer come from checking these files.

Changelog: [errors] We rewrite how we generate types for library definitions, which makes the validation of libdef files stricter. You might see more errors as a result of the stricter validation. e.g. all unbound names in libdef files will cause an error, instead of the first reference of the unbound name.

Reviewed By: panagosg7

Differential Revision: D50280494

fbshipit-source-id: 4c2c6dc90e15b9d7870b3322985428e15ccdcf22