[flow] use LocSet instead of loc lists in Get_def_result
Summary:
The result of get-def is a list (which allows duplicates). In Signature_help, this list is matched against a singleton list. This means that a list with two identical locations would unnecessarily fail this check.

This diff converts this list to a LocSet to reliably avoid duplicates, and retain the check that happens in Signature_help.

Reviewed By: SamChou19815

Differential Revision: D50355452

fbshipit-source-id: 647032b7d583b9ce853662bb8ab99c28c41c1e84