[flow] Log timing information for reinits
Summary:
During a recheck we might decide to reinit from a saved state instead. When this
happens, we were failing to log a RECHECK event with memory/timing information.
This logging is important to understand perf impacts of reinits.

This diff adds a log entry on reinit. I added a separate REINIT event because we
don't have all the same fields as a recheck has. We schedule a recheck to follow
the reinit, so there should be a separate log entry for that.

(This behavior around reinits seems to have changed in D43265619. Before that
diff, we did include reinit timings in the RECHECK event, because we did both at
the same time.)

Reviewed By: SamChou19815

Differential Revision: D50522400

fbshipit-source-id: b308c1ede27cb4935bcc533ec41cb5ce40a9fe4a