[flow] Build PropsOf<T> type
Summary: This diff builds the `PropsOf<T>` type, which receives an element type and produces its config. We express this as a conditional type and use the existing built-in React$ElementConfig utility (rather than directly extracting the config from the abstract component, because doing so breaks in the presence of default props).

Reviewed By: SamChou19815

Differential Revision: D50518466

fbshipit-source-id: 0873704456f17355caa3073f28b00238a093a4ad