[flow][render-types] Fix promotion for elements of unions
Summary:
When we see a `renders React.Element<Union>` we need to make sure that we preserve the UnionT structure, otherwise we fork computation while evaluating an EvalT and violate 0->1 rules for EvalT results. This leads to weird behavior, like spurious errors in the added test cases marked "OK".

The fix here adds a special case for UnionT ~> PromoteRendersRepresentationT. When we've already resolved an element and see a union we transform the members of the union and preserve the outer union structure.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D50620772

fbshipit-source-id: 370cceb2f3c666f9ae21c128073d72a4e72fb091