[flow][ide] Fix incorrect refactor to React components
Summary:
Fix the bug exposed in the previous diff. It turns out that we just didn't do any loc check when we are replacing, so we end up replacing everything.

Changelog: [ide] We fixed a bug for "refactor to react component code actions". Previously, it will incorrectly replace all jsx child with the newly refactored JSX elements

Reviewed By: jbrown215

Differential Revision: D50705161

fbshipit-source-id: 0c3fa9c90bd9e51b41fcac3111da2ba6a8469334