[flow] fix unused suppression warnings in incremental error collation
Summary:
When looking into why the `add_suppression_warnings` step would take longer than expected in incremental error collation, I noticed a bug that affects the "--include-warnings" output on a lazy server.

Note that by default we do not show warnings, so in most common uses the bug is not visible, but we do pay for some extra work that is unnecessary.

## The bug

Consider the file structure:
```
A1       B1 B3
 | \     | /
A2 A3    B2
```
Files A1, A3, B2 and B3 have USED suppressions.
File A2 has an UNUSED suppression.

We focus-check A1 and then B1. Under incremental error collation, checking B1 would introduce spurious unused suppression warnings on the branch of the file structure that is not the one actively checked (so in A1 and A3).

The reason for this is the following:

When checking B1 in incremental error collation we only uncover and collate errors from B1 (and B2). Notably, no errors from the A side would appear in the current error set. However, the suppressions we took into account where from the entire merged set (which includes the A side). Error collation uses actual errors to determine that a suppression is used. Thus, suppressions from the A side would not be matched by any of their respective errors, and so they showed up as "unused suppressions" warnings.

## The fix

we need to make sure that we only consider suppressions from the latest checked file set of B1 when collating errors from that side. This diff does this by doing this filtering of suppressions (similar to how we do it for errors) before they are passed to the error collator.

Note that due to misplaced errors, we still need to consider suppressions from the entire dependency set (instead of just the checked set). We approximate this by passing an "all_suppressions" parameter to error collation that is passed to Error_suppressions.

Changelog: [fix] in unused suppression warnings under incremental errors collation

Reviewed By: samwgoldman

Differential Revision: D50679015

fbshipit-source-id: 9f19cb6b4aa6468ba81c46c68aabbf7986d08f87