[flow] improvements in signature help for optional chaining and intersections
Summary:
## Changes

This diff improves signature help in a couple ways:
1. It fixes support for optional chaining (e.g. `f?.(|`, `g?.x(|`, `h?.x.y(|`, `i?.["blah"](|`). Before this diff, signature help did not return any result for these. Now it returns the same results to regular calls.
2. Improves behavior with intersection types. Before, Flow would only return the currently selected overload. In the case where the call had just started (so right after the `)`) we would always return the first overload. Now we return all available overloads that the user can switch between.

Note:

If a polymorphic function appears under an intersection, we do not show the instantiated version for that overload, but instead the generic one. For example in
```
declare function f(x: string): void;
declare function f<A>(x: A, y: A): void;
f(1,|
```
we will show `(x: string): void` and `(x: A, y: A): void`, instead of `(x: number, y: number): void` for the second option. This is an area with room for improvement. We could for example only show just the potentially successful overloads as the user types in more arguments to the call.

## Implementation

I found that the most concise way to add support for these features was through the "specialized_callee" mechanism (see D47180573). This mechanism provides the types for callee expressions in the typed AST. Its goal is to capture the specific overload and instantiation that the checker decided for each call.

In the case of signature-help, there were two main restrictions we had to take into account:

1. For optional chaining, we needed to filter out the voided values (these are reported in the case of the TAST).

2. We do not unwrap intersections. These are instead passed directly to signature help that displays them as is. (The TAST would unwrap and pick the right overload to show.) This restriction actually makes our lives easier since we can ignore `_ ~> CallT` constraints under speculation when gathering signature-help related data.

I decided to avoid storing signature-help information on the TAST itself. This would have required adding a field to `Call` which would introduce a lot of noise. Instead, I decided to populate a map stored in the context. The update of the map happens in `call_ast` in statement.ml which is the function that reconstructs the TAST. Thus it is rather unlikely that we fail to update the mapping.

Changelog: [fix] improve signature help support for optional chaining calls and intersection types

Reviewed By: SamChou19815

Differential Revision: D50360054

fbshipit-source-id: aed56b3c1f0b9e5c03f6419510e77c2102ce6870