[flow][auto-complete] Identifiers in type annotation object keys should not be autocompleted
Summary:
For object keys in value namespace, we can provide reasonable completion, since it can be used as object property shorthand. However, in type namespace, it's not a reference to any defined value in type or value, so we should not autocomplete anything at all. This would prevent some spurious auto imports going into the results.

Changelog: [ide] We no longer suggest defined values when doing auto complete in object keys when we are under a type annotation.

Reviewed By: mroch

Differential Revision: D50757976

fbshipit-source-id: daf5c0418857b0074963dab11f0af5cdb8eff222