[flow][normalizer] Print react element shorthand as `element of component Foo` in hover type and autocomplete
Summary:
In typed AST for type references, we store the type as if it's read under the value namespace. e.g.

```
class Foo {}
type A = Foo; // <- stored type is ClassT (InstanceT ...)
```

As a result, in type-at-pos and autocomplete, we print `class Foo`. I guess this is fine, since TS also does it, and people are generally don't need education that a value-namespace read and type-namespace read of a class name is different.

However, I believe this is problematic for react element shorthand. People will need education, ideally right in IDE, that the type namespace reference of `Foo` for `component Foo` means `element of component Foo`. Right now, we will print `component Foo` for the reason described above.

This diff tries to address the issue by propagating the fact that a type-at-pos query is at a type identifier reference location. If so, we will run the Ty.elt result through a function that recovers the meaning. I only did it for nominal components for now.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D50806655

fbshipit-source-id: f486cc149c7cea204369f109105f4b370ff7c685