[flow] Remove unparsed and deleted from merge context mutator
Summary:
Before D35945164, we needed to include these files to ensure that certain heap
entries were properly deleted. See D17537927 which added unparsed for this
reason.

These days, the merge context mutator only needs the set of files for rollback,
not commit, and it only rolls back changes to the values changed on merged files
(leader and sig hash).

Reviewed By: SamChou19815

Differential Revision: D50866267

fbshipit-source-id: 6cba9a8674238eebc0b46a529bb660e6440d8bc7