[flow] Pass counts to FlowEventLogger, not sets
Summary:
The event logger only needs the count, not the actual set. By pushing this up
the caller stack, we can see that we actually already calculated the count in
the caller, so we can just reuse that instead of re-computing it. (Recall that
the cardinal function in Set is O(n), not constant.)

Furthermore, and more relevant to my ongoing work, depending on the count
instead of the set itself opens the door to changing the underlying
representation.

A common complaint about this code is the complexity of having various different
sets of files with often unclear semantics.

Reviewed By: SamChou19815

Differential Revision: D50908935

fbshipit-source-id: 7b7d6d4cd77721b3036534160143196d92087f3c