[flow] Clean up unused sig_dependent_files
Summary:
With the previous diff, we no longer need to thread this set around, since we
only use all_dependent_files.

I decided to just remove the logging for the size of this set. This number
isn't used in any of our logging, and we should be using the number of all
dependent files anyway.

Reviewed By: panagosg7

Differential Revision: D50871991

fbshipit-source-id: d9e0ba971d3037e7275820f032e9e20ae8d6073a