[flow] Fix calculation of `to_merge` file set
Summary:
Background:

If we have files A and B, such that file B depends on A, B is a "signature
dependent" if the signature of B depends on A. However, if B only mentions A
within the body of a function, then it's an "implementation dependent."

Distinguishing between signature and implementation dependents allows us to
prune the set of transitive dependent files of changed files, eliminating from
consideration files which will certainly not be affected.

For example, if we have files A, B, and C, such that B implementation-depends on
A, and C implementation-depends on B, we only need to check B when A changes. To
see why, realize that the change in A can not possibly change the signature of B.

So, during a recheck, we compute the set of dependent files in two steps:

  1. Compute the transitive closure of the reverse signature dependency graph,
      which is stored in a file set called `sig_dependent_files`.

  2. Extend (1) with the direct implementation dependents of each file in (1),
      which we call `all_dependent_files`.

The include_dependents_and_dependencies function creates a number of CheckedSet
values, each containing sets of focused, dependents, and dependencies. We also
compute a list of strongly-connected components used by the merge stream to
merge files.

This diff:

Before this diff, the set `to_merge` contained the input changes, signature
dependent files, as well as dependencies which might need to be merged for a
lazy server.

However, in practice we actually merge ALL dependent files, not just signature
dependents. In fact, because we add leader information during merge, we must
merge all files, because check requires leader information.

How do we merge files which aren't in `to_merge`? Well, that set doesn't even
control what is merged. The merge stream uses `components` to drive the merge
phase, and `components` does contain all dependent files.

Changing the calulation of `to_merge` brings this set in line with what is
actually happening.

(In theory, we could move leader stuff into a different phase and avoid merging
these implementation-dependents-of-transitive-signature-dependents. However,
the complexity cost of precision here is not worth it, considering that merging
files is extremely cheap.)

Rest of stack:

I've made a surgical change here to make it easier for reviewers to understand,
but this change means a few things are now dead:

1. We don't need to separately track sig_dependent_files at all.
2. There is no difference between `to_merge` and `to_merge_or_check`

Reviewed By: panagosg7

Differential Revision: D50871992

fbshipit-source-id: 33c343d4f350d415e7ba8b9be17852ee70fe2806